requirements.txt
BeautifulSoup==3.2.1
Flask==0.10.1
Flask-Admin==1.0.7
Flask-Assets==0.8
Flask-Login==0.2.10
Flask-Mako==0.3
Flask-SQLAlchemy==1.0
Flask-Testing==0.4
Flask-WTF==0.9.1
Flask-fillin==0.2
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.18
MySQL-python==1.2.5
Pillow==2.3.0
PyHAML==1.0.1
PyYAML==3.10
SQLAlchemy==0.9.1
WTForms==1.0.5
WebHelpers==1.3
Werkzeug==0.9.4
beautifulsoup4==4.3.2
blinker==1.3
cssselect==0.9.1
extras==0.0.3
fixture==1.5
gunicorn==18.0
itsdangerous==0.23
lxml==3.3.0
mock==1.0.1
newrelic==2.6.0.5
newspaper==0.0.6
nltk==2.0.4
nose==1.3.0
passlib==1.6.2
pyScss==1.2.0.post3
python-dateutil==1.5
python-mimeparse==0.1.4
requests==1.2.3
six==1.5.2
tld==0.6.3
twill==0.9
webassets==0.8
wsgiref==0.1.2
